split out universal route handler
