+++ src/server/routes/index.js
@@ -1,0 +1,9 @@
+import login from './login';
+import api from './api';
+import universal from './universal';
+
+export {
+  login,
+  api,
+  universal
+};
