+++ src/server/routes/universal.js
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+import { universal } from '../handlers/universal';
+
+const router = Router();
+
+router.get('*', universal);
+
+export default router;
