+++ src/server/server.js
@@ -1,21 +1,12 @@
 import Express from 'express';
-import React from 'react';
 import fs from 'fs';
 import helmet from 'helmet';
 import morgan from 'morgan';
 import path from 'path';
 import session from 'express-session';
-import url from 'url';
-import util from 'util';
-import { match, RouterContext } from 'react-router';
-import { renderToString } from 'react-dom/server';
 
-import Html from '../server/helpers/html';
-import api from './routes/api';
-import conf from './configure.js';
-import configureStore from '../common/store/configureStore';
-import login from './routes/login';
-import routes from '../common/routes';
+import * as routes from './routes/';
+import conf from './configure';
 import sessionConfig from './helpers/session';
 
 const logsPath = conf.get('SERVER:LOGS_PATH') || path.join(__dirname, '../../logs/');
@@ -35,80 +26,10 @@ if (app.get('env') === 'production') {
   app.set('trust proxy', 1);
 }
 
-app.use('/', login);
-app.use('/api', api);
-
 app.use(Express.static(__dirname + '/../public'));
-app.use('/', function (req, res) {
 
-
-  match({ routes, location: req.url }, (error, redirectLocation, renderProps) => {
-    if (error) {
-      res.status(500).send(error.message);
-    } else if (redirectLocation) {
-      res.redirect(302, redirectLocation.pathname + redirectLocation.search);
-    } else if (renderProps) {
-
-      const initialState = {};
-
-      if (req.session) {
-
-        if (req.session.authenticated) {
-          initialState['identity'] = {
-            isAuthenticated: req.session.authenticated,
-            name: req.session.name,
-            email: req.session.email
-          };
-        }
-
-
-        if (req.session.error) {
-          initialState['oyez'] = [{
-            'message': req.session.error,
-            'status': 'error'
-          }];
-
-          delete req.session.error;
-        }
-
-      }
-
-
-      const store = configureStore(initialState);
-
-      // config we share from server side to client side
-      const config = {
-        UNIVERSAL: conf.get('UNIVERSAL')
-      };
-
-      // You can also check renderProps.components or renderProps.routes for
-      // your "not found" component or route respectively, and send a 404 as
-      // below, if you're using a catch-all route.
-      const component = <RouterContext {...renderProps} />;
-
-      res.status(200);
-      res.send('<!doctype html>\n' +
-        renderToString(
-          <Html
-            store={ store }
-            component={ component }
-            config={ config }
-          />
-        ));
-    } else {
-      res.status(404).send('Not found');
-    }
-  });
-});
-
-const appUrl = url.parse(conf.get('UNIVERSAL:MU_URL'));
-const appHost = conf.get('SERVER:HOST') || appUrl.hostname;
-const appPort = conf.get('SERVER:PORT') || appUrl.port;
-const server = app.listen(appPort, appHost, () => {
-  const host = server.address().address;
-  const port = server.address().port;
-
-  util.log('🚂  Express server listening on http://%s:%s 🚂', host, port);
-});
+app.use('/', routes.login);
+app.use('/api', routes.api);
+app.use('/', routes.universal);
 
 export { app as default };
